patched structures mismatch with winStructs causing the service to crash. number of attempts to patch = 7, number of hair lost = i lost count

+++ EtwStructure/RealTimeConsumer/Native/Structures/BufferContextStruct.cs
@@ -6,8 +6,8 @@ namespace EtwIpGrabber.EtwStructure.RealTimeConsumer.Native.Structures
     [StructLayout(LayoutKind.Sequential)]
     public struct ETW_BUFFER_CONTEXT
     {
-        public ushort ProcessorNumber;
-        public ushort Alignment;
+        public byte ProcessorNumber;
+        public byte Alignment;
         public ushort LoggerId;
     }
 }
